OCSADV-417 Refactor of the Catalog Navigator UI separating classes into their own files to reduce complexity #time 1h
