+++ bundle/jsky.app.ot/src/main/java/jsky/app/ot/gemini/editor/targetComponent/EdCompTargetList.java
@@ -1,6 +1,6 @@
 package jsky.app.ot.gemini.editor.targetComponent;
 
-import edu.gemini.catalog.ui.QueryResultsWindow;
+import edu.gemini.catalog.ui.QueryResultsFrame;
 import edu.gemini.pot.sp.ISPObsComponent;
 import edu.gemini.shared.skyobject.Magnitude;
 import edu.gemini.shared.util.immutable.*;
@@ -69,8 +69,10 @@ public final class EdCompTargetList extends OtItemEditor<ISPObsComponent, Target
         _w.primaryButton  .addActionListener(primaryListener);
 
         _w.guidingControls.manualGuideStarButton().peer().addActionListener(manualGuideStarListener);
-        _w.guidingControls.newManualGuideStarButton().peer().addActionListener(evt ->
-                QueryResultsWindow.instance().showOn(getNode()));
+
+        _w.guidingControls.newManualGuideStarButton().peer().addActionListener(evt -> {
+            QueryResultsFrame.instance().showOn(getNode());
+        });
         _w.guidingControls.autoGuideStarGuiderSelector().addSelectionListener(strategy ->
                 AgsStrategyUtil.setSelection(getContextObservation(), strategy)
         );
